+++ src/Api/Representation/TaxonomyTermRepresentation.php
@@ -85,7 +85,7 @@ class TaxonomyTermRepresentation extends AbstractResourceEntityRepresentation
         );
     }
 
-    public function subjectValueTotalCount($property = null)
+    public function subjectValueTotalCount($property = null, $resourceType = null, $siteId = null)
     {
         return $this->getAdapter()->getSubjectValueTotalCount($this->resource, $property);
     }